legal_ai/requirements.txt
# Legal AI - Turkish Criminal Law RAG
# Python 3.11

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# RAG & LLM
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langchain-text-splitters>=0.0.1

# Embeddings & vector store
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4

# PDF loading
pypdf>=3.17.0

# Validation & types
pydantic>=2.5.0

# HuggingFace for bge-m3
huggingface-hub>=0.20.0
